Refactor: Add dir /editor/toolbar
